Added the FirebaseMessagesRepository class
